Move skeletons and fix pub keys
