+++ src/app/[locale]/(logged-out)/data-custodian-network/[dataCustodianNetworkId]/components/Skeletons/index.ts
@@ -1,0 +1,3 @@
+import { NetworkSkeleton, SectionSkeleton } from "./Skeletons";
+
+export { NetworkSkeleton, SectionSkeleton };
